feat(CopyPaste): Use swift script to get all files path from clipboard (WIP)
